alert

+++ app/src/main/java/pt/utl/ist/airdesk/airdesk/ViewWorkspace.java
@@ -3,6 +3,7 @@ package pt.utl.ist.airdesk.airdesk;
 import android.app.AlertDialog;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.os.AsyncTask;
 import android.os.Environment;
 import android.support.v7.app.ActionBarActivity;
 import android.os.Bundle;
@@ -18,15 +19,28 @@ import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.io.ObjectInputStream;
+import java.io.ObjectOutputStream;
 import java.util.ArrayList;
+import java.util.Arrays;
 
+import pt.inesc.termite.wifidirect.sockets.SimWifiP2pSocket;
 import pt.utl.ist.airdesk.airdesk.Sqlite.UsersDataSource;
 import pt.utl.ist.airdesk.airdesk.Sqlite.WSDataSource;
 import pt.utl.ist.airdesk.airdesk.Sqlite.WSPermissionSource;
 import pt.utl.ist.airdesk.airdesk.datastructures.DeviceInformation;
+import pt.utl.ist.airdesk.airdesk.datastructures.FileDeleteRequest;
+import pt.utl.ist.airdesk.airdesk.datastructures.FileDeleteResponse;
+import pt.utl.ist.airdesk.airdesk.datastructures.FileLockRequest;
+import pt.utl.ist.airdesk.airdesk.datastructures.FileLockResponse;
+import pt.utl.ist.airdesk.airdesk.datastructures.FileResponse;
+import pt.utl.ist.airdesk.airdesk.datastructures.FileResponseAlteration;
+import pt.utl.ist.airdesk.airdesk.datastructures.WorkspaceRepToBeSent;
+import pt.utl.ist.airdesk.airdesk.datastructures.WorkspacesShared;
 
 
 public class ViewWorkspace extends ActionBarActivity {
@@ -133,18 +147,63 @@ public class ViewWorkspace extends ActionBarActivity {
 
                 alert.setPositiveButton("yes", new DialogInterface.OnClickListener() {
                     public void onClick(DialogInterface dialog, int whichButton) {
-                    File fileToDelete = new File(path,filesList.get(position).toString());
-                        fileToDelete.delete();
-                        filesList.clear();
-                        File file[] = f.listFiles();
-                        if(!(file == null)) {
-                            for (int i = 0; i < file.length; i++) {
-                                filesList.add(file[i].getName());
+                        if (ambiente.equals("local")) {
+                            File fileToDelete = new File(path, filesList.get(position).toString());
+                            fileToDelete.delete();
+                            filesList.clear();
+                            File file[] = f.listFiles();
+                            if (!(file == null)) {
+                                for (int i = 0; i < file.length; i++) {
+                                    filesList.add(file[i].getName());
+                                }
                             }
-                        }
-                        listAdapter.notifyDataSetChanged();
-                        dialog.dismiss();
+                            listAdapter.notifyDataSetChanged();
+                            dialog.dismiss();
 
+                        } else {
+
+                            final File fileToDelete = new File(path, filesList.get(position).toString());
+                            Thread t = new Thread() {
+                                public void run() {
+
+                                    try {
+
+                                        Log.v("conadamae", "antes");
+
+                                        FileDeleteRequest fileDeleteRequest = new FileDeleteRequest(fileToDelete,login,workspace);
+
+
+                                        final SimWifiP2pSocket s = new SimWifiP2pSocket(deviceInformation.getIp(), deviceInformation.getPort());
+
+                                        ObjectOutputStream oos = new ObjectOutputStream(s.getOutputStream());
+                                        oos.writeObject(fileDeleteRequest);
+
+                                        runOnUiThread(new Runnable() {
+                                            @Override
+                                            public void run() {
+
+                                                new ReceiveFileCommTask().executeOnExecutor(
+                                                        AsyncTask.THREAD_POOL_EXECUTOR, s);
+                                            }
+                                        });
+
+
+
+                                        //mCliSocket.getOutputStream().write( ("hello world" + "\n").getBytes());
+
+                                        Log.v("conadamae", "depois");
+
+                                        //mCliSocket.getOutputStream().write( ("hello world" + "\n").getBytes());
+
+                                        Log.v("conadamae", "depois");
+                                    } catch (IOException e) {
+                                        e.printStackTrace();
+                                    }
+
+                                }
+                            };
+                            t.start();
+                        }
                     }
                 });
 
@@ -205,31 +264,60 @@ else {
 
             alert.setTitle("Invite user");
 
+
             alert.setMultiChoiceItems(items, null, new DialogInterface.OnMultiChoiceClickListener() {
                 @Override
                 public void onClick(DialogInterface dialog, int which, boolean isChecked) {
-                    if(isChecked){
+                    if (isChecked) {
                         selectedItems.add(which);
-                    }
-                    else if (selectedItems.contains(which)){
+
+
+                    } else if (selectedItems.contains(which)) {
                         selectedItems.remove(Integer.valueOf(which));
                     }
                 }
             });
 
-            alert.setMessage("Name of the user");
+            //alert.setMessage("Name of the user");
 
 
             final EditText input = new EditText(this);
+            input.setHint("Insert username to share here");
             alert.setView(input);
 
             alert.setPositiveButton("yes", new DialogInterface.OnClickListener() {
                 public void onClick(DialogInterface dialog, int whichButton) {
 
-                   datasourcePermissions.createPermissionsRepresentation(workspace,input.getText().toString(),"rw" );
-                    Toast.makeText(getApplicationContext(), "workspace shared with " + input.getText().toString(),
-                            Toast.LENGTH_LONG).show();
+                    StringBuilder stringBuilder = new StringBuilder();
 
+
+                    if (selectedItems.contains(0)) {
+                        //permission = "rw_";
+                        stringBuilder.append("rw_");
+                    }
+                    if (selectedItems.contains(1)) {
+                        stringBuilder.append("c_");
+                    }
+                    if (selectedItems.contains(2)) {
+                        stringBuilder.append("d_");
+                    }
+
+
+                    permission = stringBuilder.toString();
+
+                    Log.v("conadamae", permission);
+
+                    if(datasourcePermissions.getPermission(workspace,input.getText().toString()).equals("Not_Found")) {
+
+                        datasourcePermissions.createPermissionsRepresentation(workspace, input.getText().toString(), permission);
+                        Toast.makeText(getApplicationContext(), "workspace shared with " + input.getText().toString(),
+                                Toast.LENGTH_LONG).show();
+                        Toast.makeText(getApplicationContext(), "Permission " + permission,
+                                Toast.LENGTH_LONG).show();
+                    }
+                    else {
+                        datasourcePermissions.updatePermission(input.getText().toString(),workspace,permission);
+                    }
                 }
             });
 
@@ -280,5 +368,82 @@ else {
         }
 
         return super.onOptionsItemSelected(item);
+    }
+
+    public class ReceiveFileCommTask extends AsyncTask<SimWifiP2pSocket, String, Void> {
+        SimWifiP2pSocket s;
+
+        @Override
+        protected Void doInBackground(SimWifiP2pSocket... params) {
+            BufferedReader sockIn;
+            String st;
+
+            s = params[0];
+            try {
+                while(true){
+                    // sockIn = new BufferedReader(new InputStreamReader(s.getInputStream()));
+                    Log.v("conadamae", "receivecomm");
+
+                    ObjectInputStream ois = new ObjectInputStream(s.getInputStream());
+
+                    ArrayList<WorkspaceRepToBeSent> lwrtbs = new ArrayList<WorkspaceRepToBeSent>();
+                    WorkspacesShared foreignWsReceived;
+                    Object o = ois.readObject();
+
+
+                    if (o instanceof FileDeleteResponse){
+                        FileDeleteResponse fileDeleteResponse;
+                        fileDeleteResponse = (FileDeleteResponse) o;
+                        publishProgress(fileDeleteResponse.getStatus(),"FileDelete",fileDeleteResponse.getFileName());
+                        break;
+                    }
+
+
+                    Log.v("conadamae", "recebi");
+                }
+                //   while ((st = sockIn.readLine()) != null) {
+                //       publishProgress(st);
+
+                // Toast.makeText(getApplicationContext(),"recebi: " + st, Toast.LENGTH_LONG).show();
+
+            } catch (IOException e) {
+                Log.d("Error reading socket:", e.getMessage());
+            }catch(ClassNotFoundException e){e.getMessage();}
+            return null;
+        }
+
+        @Override
+        protected void onPreExecute() {
+
+
+        }
+
+        @Override
+        protected void onProgressUpdate(String... values) {
+        if(values[1].equals("FileDelete")){
+            if(values[0].equals("Deleted")){
+                Toast.makeText(getApplicationContext(), "File was deleted!", Toast.LENGTH_LONG).show();
+                filesList.remove(values[2]);
+                listAdapter.notifyDataSetChanged();
+            }
+            else{
+                Toast.makeText(getApplicationContext(), "You don't have permission to delete file", Toast.LENGTH_LONG).show();
+            }
+        }
+        }
+
+        @Override
+        protected void onPostExecute(Void result) {
+            if (!s.isClosed()) {
+                try {
+                    s.close();
+                }
+                catch (Exception e) {
+                    Log.d("Error closing socket:", e.getMessage());
+                }
+            }
+            s = null;
+
+        }
     }
 }
